requirements.txt
boto3
matplotlib
pillow